requirements.txt
django==4.0.3
djangorestframework==3.13.1

django-extensions==3.1.5
djangorestframework==3.13.1
django-filter==21.1
drf-spectacular==0.22.0